% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lm_model}
\alias{glm_model}
\alias{glm_nb_model}
\alias{brm_model}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{trending}{\code{\link[trending:trending_model]{brm_model}}, \code{\link[trending:trending_model]{glm_model}}, \code{\link[trending:trending_model]{glm_nb_model}}, \code{\link[trending:trending_model]{lm_model}}}
}}

