% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aic.R, R/calculate_yardstick_metric.R
\name{calculate_aic}
\alias{calculate_aic}
\alias{calculate_aic.default}
\alias{calculate_aic.trending_fit}
\alias{calculate_aic.trending_model}
\alias{calculate_aic.list}
\title{Generic for calculating the AIC}
\usage{
calculate_aic(x, ...)

\method{calculate_aic}{default}(x, ...)

\method{calculate_aic}{trending_fit}(x, ...)

\method{calculate_aic}{list}(x, data, ...)

\method{calculate_aic}{list}(x, data, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Not currently used.}
}
\value{
For a single \code{\link[trending:reexports]{trending_fit}} input, the object
returned will be a list with entries:
\itemize{
\item metric: "AIC"
\item result: the resulting AIC value fit (NULL if the calculation failed)
\item warnings: any warnings generated during calculation
\item errors: any errors generated during calculation
}

For a list of trending fit inputs a \link[tibble:tibble]{tibble} with one
row for each fitted model and corresponding to output generated with a
single model input.
}
\description{
Generic \code{calculate_aic()} returns the Akaike's 'An Information Criterion' for
the given input.
}
\details{
Specific methods are given for \code{\link[trending:reexports]{trending_fit}}
objects and lists of these models. The default method applies
\code{\link[stats:AIC]{stats::AIC()}} directly.
}
\author{
Tim Taylor

#' @examples
x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = "poisson")
negbin_model <- glm.nb_model(y ~ x)
fitted_model <- fit(poisson_model, dat)
fitted_models <- lapply(c(poisson_model, negbin_model), fit, data = dat)

calculate_aic(poisson_model, dat)
calculate_aic(fitted_models)

calculate_aic(fitted_model)
}
